requirements.txt
Flask==3.0.3
gunicorn==22.0.0
Pillow==10.4.0
numpy==2.1.2
pandas==2.2.2
joblib==1.4.2
